add only if to workaround puppet bug #12830

+++ manifests/logrotate/centos.pp
@@ -4,5 +4,6 @@ class apache::logrotate::centos {
       changes => [ 'rm /files/etc/logrotate.d/httpd/rule/file',
         'ins file before /files/etc/logrotate.d/httpd/rule/*[1]',
         'set /files/etc/logrotate.d/httpd/rule/file[1] /var/log/httpd/*log' ],
+      onlyif => 'get /files/etc/logrotate.d/httpd/rule/file[1] != "/var/log/httpd/*log"',
     }
 }
